requirements.txt
pyyaml>=6.0.1,<7
python-dotenv>=1.0.1,<2
litellm>=1.74.0,<2
mcp>=1.0.0,<2
pyghidra>=3.0.0,<4
fasteners>=0.19,<1
requests>=2.31.0,<3
